Fixes issue #225 and #243 issue with text to speechand Map feature broken issue
Fixes issue #225 and #243 issue with text to speechand Map feature broken issue
