+++ app/src/main/java/org/fossasia/susi/ai/adapters/recyclerAdapters/ChatFeedRecyclerAdapter.java
@@ -82,8 +82,6 @@ public class ChatFeedRecyclerAdapter extends RealmRecyclerViewAdapter<ChatMessag
 //    private Activity activity;
     private String TAG = ChatFeedRecyclerAdapter.class.getSimpleName();
     private RecyclerView recyclerView;
-    private TextToSpeech textToSpeech;
-    private AudioManager.OnAudioFocusChangeListener afChangeListener;
 
     public ChatFeedRecyclerAdapter(@NonNull Context context, @Nullable OrderedRealmCollection<ChatMessage> data, boolean autoUpdate) {
         super(context, data, autoUpdate);
@@ -268,43 +266,9 @@ public class ChatFeedRecyclerAdapter extends RealmRecyclerViewAdapter<ChatMessag
                         chatViewHolder.timeStamp.setTag(chatViewHolder);
                         break;
                     case SUSI_MESSAGE:
-                        final String toSpeak = model.getContent();
                         chatViewHolder.chatTextView.setText(model.getContent());
                         chatViewHolder.timeStamp.setText(model.getTimeStamp());
                         chatViewHolder.chatTextView.setTag(chatViewHolder);
-                        if(MainActivity.checkSpeechOutputPref()){
-                            if(MainActivity.checkSpeechOutputPref()){
-                                final AudioManager audiofocus = (AudioManager) currContext.getSystemService(Context.AUDIO_SERVICE);
-                                int result = audiofocus.requestAudioFocus(afChangeListener,AudioManager.STREAM_MUSIC,AudioManager.AUDIOFOCUS_GAIN);
-                                if(result== AudioManager.AUDIOFOCUS_REQUEST_GRANTED)
-                                {
-                                    textToSpeech=new TextToSpeech(currContext, new TextToSpeech.OnInitListener() {
-                                        @Override
-                                        public void onInit(int status) {
-                                            if(status != TextToSpeech.ERROR) {
-                                                textToSpeech.setLanguage(Locale.UK);
-                                                if(position==getItemCount()-1)
-                                                    textToSpeech.speak(toSpeak, TextToSpeech.QUEUE_FLUSH, null);
-                                                audiofocus.abandonAudioFocus(afChangeListener);
-
-                                            }
-                                        }
-                                    });}
-
-                                AudioManager.OnAudioFocusChangeListener afChangeListener =
-                                        new AudioManager.OnAudioFocusChangeListener() {
-                                            public void onAudioFocusChange(int focusChange) {
-                                                if (focusChange == AUDIOFOCUS_LOSS_TRANSIENT) {
-                                                    textToSpeech.stop();
-                                                } else if (focusChange == AudioManager.AUDIOFOCUS_GAIN) {
-                                                    // Resume playback
-                                                } else if (focusChange == AudioManager.AUDIOFOCUS_LOSS) {
-                                                    textToSpeech.stop();
-                                                }
-                                            }
-                                        };}}
-
-
                         chatViewHolder.chatMessage.setOnLongClickListener(new View.OnLongClickListener() {
                             @Override
                             public boolean onLongClick(final View view) {
@@ -363,39 +327,6 @@ public class ChatFeedRecyclerAdapter extends RealmRecyclerViewAdapter<ChatMessag
     private void handleItemEvents(final MapViewHolder mapViewHolder, final int position) {
 
         final ChatMessage model = getData().get(position);
-        final String toSpeak = model.getContent();
-        String[] parts = toSpeak.split(":");
-        final String string1 = parts[0];
-        if(MainActivity.checkSpeechOutputPref()){
-            final AudioManager audiofocus = (AudioManager) currContext.getSystemService(Context.AUDIO_SERVICE);
-            int result = audiofocus.requestAudioFocus(afChangeListener,AudioManager.STREAM_MUSIC,AudioManager.AUDIOFOCUS_GAIN);
-            if(result== AudioManager.AUDIOFOCUS_REQUEST_GRANTED)
-            {
-                textToSpeech=new TextToSpeech(currContext, new TextToSpeech.OnInitListener() {
-                    @Override
-                    public void onInit(int status) {
-                        if(status != TextToSpeech.ERROR) {
-                            textToSpeech.setLanguage(Locale.UK);
-                            if(position==getItemCount()-1)
-                                textToSpeech.speak(string1, TextToSpeech.QUEUE_FLUSH, null);
-                            audiofocus.abandonAudioFocus(afChangeListener);
-
-                        }
-                    }
-                });}
-
-            AudioManager.OnAudioFocusChangeListener afChangeListener =
-                    new AudioManager.OnAudioFocusChangeListener() {
-                        public void onAudioFocusChange(int focusChange) {
-                            if (focusChange == AUDIOFOCUS_LOSS_TRANSIENT) {
-                                textToSpeech.stop();
-                            } else if (focusChange == AudioManager.AUDIOFOCUS_GAIN) {
-
-                            } else if (focusChange == AudioManager.AUDIOFOCUS_LOSS) {
-                                textToSpeech.stop();
-                            }
-                        }
-                    };
         mapViewHolder.chatMessages.setOnLongClickListener(new View.OnLongClickListener() {
             @Override
             public boolean onLongClick(final View view) {
@@ -467,7 +398,7 @@ public class ChatFeedRecyclerAdapter extends RealmRecyclerViewAdapter<ChatMessag
                 e.printStackTrace();
             }
         }
-    }}
+    }
 
     private void deleteMessage(final int position) {
         realm.executeTransaction(new Realm.Transaction() {
